Fixed issue with removing items from an array
Polymer uses WeakMap for collections. As a result of this, when
listening for data binding changes on an array, the path passed to the
`value-changed` event listener uses the id for the object in the map,
not the index. As Polymer uses the index of the last item added as the
key for the WeakMap, it appears the path constructed for
`value-changed` events is the index of the item in the array, until an
item is removed with `.splice()`. In order to correctly create `value`
for array properties, the association between array item index and
WeakMap key (used to construct path) must be tracked.